Move MockFileDistributionFactory to another packge as well
