+++ configserver/src/main/java/com/yahoo/vespa/serviceview/Cluster.java
@@ -9,7 +9,7 @@ import com.google.common.collect.ImmutableList;
 /**
  * Model a single cluster of services in the Vespa model.
  *
- * @author <a href="mailto:steinar@yahoo-inc.com">Steinar Knutsen</a>
+ * @author Steinar Knutsen
  */
 public final class Cluster implements Comparable<Cluster> {
